Refactor: Move TextEditor handling in its own component
Preparations for using the new editor API from Text, see
https://github.com/nextcloud/text/pull/3615
